[change] changed type from 'double' to 'float'

+++ obj.cpp
@@ -4,7 +4,7 @@ Vec3::Vec3() : Vec3(0.0, 0.0, 0.0)
 {
 }
 
-Vec3::Vec3(double x, double y, double z)
+Vec3::Vec3(float x, float y, float z)
 {
     this->x = x;
     this->y = y;
@@ -15,7 +15,7 @@ Vec3::~Vec3()
 {
 }
 
-double Vec3::length()
+float Vec3::length()
 {
     return sqrt(std::pow(this->x, 2) + std::pow(this->y, 2) + std::pow(this->z, 2));
 }
@@ -38,10 +38,10 @@ Face::Face(Vec3 p1, Vec3 p2, Vec3 p3, Vec3 normal)
     this->normal = normal;
 }
 
-Face::Face(double x1, double y1, double z1,
-           double x2, double y2, double z2,
-           double x3, double y3, double z3,
-           double nx, double ny, double nz) : Face(Vec3(x1, y1, z1), Vec3(x2, y2, z2), Vec3(x3, y3, z3), Vec3(nx, ny, nz))
+Face::Face(float x1, float y1, float z1,
+           float x2, float y2, float z2,
+           float x3, float y3, float z3,
+           float nx, float ny, float nz) : Face(Vec3(x1, y1, z1), Vec3(x2, y2, z2), Vec3(x3, y3, z3), Vec3(nx, ny, nz))
 {
 }
